refactor: badge module to custom forms module
