--- src/app/shared/modules/badge/badge.component.html
@@ -1,8 +1,0 @@
-
-<div class="flex">
-  <span
-    *ngFor="let badge of badges"
-    class="inline-flex items-center justify-center px-2 py-1 mr-2 text-xs font-bold leading-none text-red-100 bg-red-600 rounded-full">
-      {{ badge }}
-  </span>
-</div>
