--- src/app/shared/modules/badge/badge.component.ts
@@ -1,17 +1,0 @@
-import { Component, Input, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-badge',
-  templateUrl: './badge.component.html',
-  styleUrls: ['./badge.component.scss'],
-})
-export class BadgeComponent implements OnInit {
-
-  @Input() badges: string[] = [];
-
-
-  constructor() { }
-
-  ngOnInit(): void {}
-
-}
