+++ src/app/shared/modules/custom-forms/badge/badge.component.html
@@ -1,0 +1,19 @@
+<!--
+<div class="flex">
+  <span
+    *ngFor="let badge of badges"
+    class="inline-flex items-center justify-center px-2 py-1 mr-2 text-xs font-bold leading-none text-red-100 bg-red-600 rounded-full">
+      {{ badge }}
+  </span>
+</div> -->
+
+<ng-container *ngFor="let badge of badges; let i = index">
+  <span
+    (click)="clickBadge(i)"
+    class="inline-flex items-center justify-center px-2 py-1 mr-2 text-xs font-bold leading-none text-red-100 bg-red-600 rounded-full cursor-pointer">
+      {{ badge }}
+      <svg *ngIf="deleteOption" xmlns="http://www.w3.org/2000/svg" class="h-4 w-4" fill="none" viewBox="0 0 24 24" stroke="currentColor">
+        <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M6 18L18 6M6 6l12 12" />
+      </svg>
+  </span>
+</ng-container>
