+++ src/app/shared/modules/custom-forms/badge/interfaces/card.interface.ts
