+++ src/app/shared/modules/custom-forms/custom-forms.module.ts
@@ -3,10 +3,12 @@ import { CommonModule } from '@angular/common';
 import { IonicModule } from '@ionic/angular';
 import { FormsAddBadgeComponent } from './components/forms-add-badge/forms-add-badge.component';
 import { ReactiveFormsModule } from '@angular/forms';
+import { BadgeComponent } from './badge/badge.component';
 
 
 @NgModule({
   declarations: [
+    BadgeComponent,
     FormsAddBadgeComponent
   ],
   imports: [
@@ -15,6 +17,7 @@ import { ReactiveFormsModule } from '@angular/forms';
     ReactiveFormsModule
   ],
   exports: [
+    BadgeComponent,
     FormsAddBadgeComponent
   ]
 })
